backendberry/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
base==0.0.0
beautifulsoup4==4.12.2
blinker==1.6.3
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.3
contextlib2==21.6.0
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.1
distro==1.8.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
effdet==0.4.1
emoji==2.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
et-xmlfile==1.1.0
exceptiongroup==1.2.0
execnet==2.0.2
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.324
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.52
layoutparser==0.3.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
mock==5.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2
nltk==3.8.1
node==1.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
odict==1.9.0
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==0.28.1
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
path==16.9.0
path.py==12.5.0
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pluggy==1.3.0
plumber==1.7
portalocker==2.8.2
preshed==3.0.9
protobuf==4.24.4
py==1.4.22
pyarrow==13.0.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.22.0
pytesseract==0.3.10
pytest==2.6.0
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
sqlparse==0.4.4
srsly==2.4.8
streamlit==1.28.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.5.1
timm==0.9.8
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uno==0.3.3
unoconv==0.9.0
unstructured==0.10.27
unstructured-inference==0.7.10
unstructured.pytesseract==0.3.12
urllib3==2.0.7
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
weasel==0.3.3
weaviate-client==3.24.2
XlsxWriter==3.1.9
yarl==1.9.2
zipp==3.17.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.1
zope.lifecycleevent==5.0
zope.proxy==5.1
